Move all fixtures from train fixtures to ndbfixtures train
